+++ packages/docs-infra/src/pipeline/loadServerTypesText/loadServerTypesText.test.ts
@@ -2,8 +2,9 @@ import { describe, it, expect, afterEach } from 'vitest';
 import { access, mkdir, readFile, rm, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 import { pathToFileURL } from 'node:url';
-import { loadServerTypesText, normalizeTypesSourceDataForCache } from './loadServerTypesText';
+import { loadServerTypesText } from './loadServerTypesText';
 import type { TypesSourceData } from './loadServerTypesText';
+import { normalizeTypesSourceDataForCache } from './normalizeTypesSourceDataForCache';
 import { buildTypesTextCacheContent } from './resolveTypesCacheKey';
 import { hashCacheContent, resolveCachePath, saveFileCache } from '../cacheUtils';
 import { generateTypesMarkdown } from '../syncTypes/generateTypesMarkdown';
